+++ dotnet/src/dotnetframework/GxClasses/Helpers/HttpHelper.cs
@@ -25,7 +25,6 @@ using System.Web;
 using System.Runtime.Serialization;
 using GeneXus.Mime;
 using System.Text.RegularExpressions;
-using System.Net.Http.Headers;
 
 namespace GeneXus.Http
 {
@@ -351,39 +350,6 @@ namespace GeneXus.Http
 			}
 		}
 
-		public static string GetEncodedContentDisposition(string value, int browserType)
-		{
-			int filenameIdx = value.IndexOf("filename", StringComparison.OrdinalIgnoreCase);
-			int eqIdx = value.IndexOf("=", filenameIdx);
-			if (filenameIdx == -1 || eqIdx == -1 || browserType == GxContext.BROWSER_SAFARI) //Safari does not supports yet ContentDispositon file name encoding value.
-			{
-				return value;
-			}
-
-			string rawFilename = value.Substring(eqIdx + 1).Trim();
-			try
-			{
-				string dispositionType = value.Substring(0, value.IndexOf(";")).Trim();
-				value = new ContentDispositionHeaderValue(dispositionType) { FileName = rawFilename }.ToString();
-			}
-			catch (Exception)
-			{
-				value = value.Substring(0, eqIdx + 1) + EncodeContentDispositionFileName(rawFilename);
-			}
-			return value;
-		}
-
-		private static string EncodeContentDispositionFileName(string filename)
-		{
-			try
-			{
-				return Uri.EscapeDataString(filename);
-			}
-			catch (UriFormatException) //Contains High Surrogate Chars
-			{
-				return GXUtil.UrlEncode(filename);
-			}
-		}
 	}
 #if NETCORE
 	public class HttpCookieCollection : Dictionary<string, HttpCookie>
