+++ dotnet/test/DotNetCoreUnitTest/HttpUtils/HttpUtils.cs
@@ -1,6 +1,6 @@
 using System;
 using GeneXus.Application;
-using GeneXus.Http;
+using GeneXus.Utils;
 using Xunit;
 
 namespace DotNetCoreUnitTest.HttpUtils
@@ -58,7 +58,7 @@ namespace DotNetCoreUnitTest.HttpUtils
 
 		private static void DoTest(string contentDisposition, string expectedContentDisposition, int browserType = GxContext.BROWSER_CHROME)
 		{
-			string encodedValue = HttpHelper.GetEncodedContentDisposition(contentDisposition, browserType);
+			string encodedValue = GXUtil.EncodeContentDispositionHeader(contentDisposition, browserType);
 
 			Assert.Equal(expectedContentDisposition, encodedValue);
 		}
